source/cdk/canary-token-lambda/requirements.txt
pyjwt
boto3